refactor api config
