+++ composer.json
@@ -5,8 +5,13 @@
     },
     "autoload": {
         "files": [
-            "./logger.php"
-       
+            "./logger.php",
+            "./authen_author.php",
+            "./modules/shoe/shoe_manage.php",
+            "./modules/car/car_manage.php",
+            "./modules/role/role_manage.php",
+            "./Utils.php",
+            "./api_router.php"
         ]
     }
 }
